tes.py
import warnings
warnings.filterwarnings("ignore")
import streamlit as st
import tempfile
import av
import os
import tensorflow as tf
from object_detection.utils import label_map_util
from object_detection.utils import visualization_utils as viz_utils
from object_detection.builders import model_builder
from object_detection.utils import config_util