requirements.txt
avro-python3
apache-beam
pillow
lxml
matplotlib
Cython
contextlib2
tf-slim
six
pycocotools
lvis
scipy
pandas
tf-models-official>=2.5.1
tensorflow_io
keras
pyparsing==2.4.7
sacrebleu<=2.2.0
object_detection==0.0.3
easyocr
streamlit
av==10.0.0
wget